Journal Componenet Rename
