+++ src/pages/journal/journal-components/indicators/indicators.html
