+++ src/pages/journal/journal-components/indicators/indicators.scss
@@ -1,0 +1,11 @@
+indicators {
+  div {
+    display: flex;
+    justify-content: center;
+    width: 2em;
+    img {
+      width: 1.8em;
+      height: 1.8em;
+    }
+  }
+}
