+++ src/pages/journal/journal-components/indicators/indicators.ts
@@ -1,0 +1,12 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'indicators',
+  templateUrl: 'indicators.html'
+})
+export class IndicatorsComponent {
+
+  @Input()
+  showSpecialNeedsIndicator: boolean;
+
+}
