--- src/pages/journal/journal-components/journal-candidate/journal-candidate.scss
@@ -1,29 +1,0 @@
-journal-candidate {
-    & > div {
-      height: 100%;
-      padding: 0.3em 0.75em;
-      background-color: map-get($colors-gds, gds-grey-4);
-      border-radius: 13px;
-      @include box-shadow(0px, 2px, 4px, rgba(0, 0, 0, 0.55));
-      &.journal-candidate-test-complete-text {
-        color: map-get($colors-gds, gds-grey-1);
-      }
-      img {
-        &.journal-candidate-category-icon-completed {
-          opacity: 0.5;
-        }
-      }
-      .catb {
-        width: 35px;
-        height: 17px;
-      }
-      .journal-candidate-name {
-        font-weight: bold;
-        overflow: hidden;
-        text-overflow: ellipsis;
-      }
-      .journal-test-category-category-complete-text {
-        color: map-get($colors, journalTestCategory )
-      }
-    }
-  }
