--- src/pages/journal/journal-components/journal-indicators/journal-indicators.scss
@@ -1,11 +1,0 @@
-journal-indicators {
-  div {
-    display: flex;
-    justify-content: center;
-    width: 2em;
-    img {
-      width: 1.8em;
-      height: 1.8em;
-    }
-  }
-}
