--- src/pages/journal/journal-components/journal-indicators/journal-indicators.ts
@@ -1,12 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-  selector: 'journal-indicators',
-  templateUrl: 'journal-indicators.html'
-})
-export class JournalIndicatorsComponent {
-
-  @Input()
-  showSpecialNeedsIndicator: boolean;
-
-}
