--- src/pages/journal/journal-components/journal-slot/journal-slot.html
@@ -1,23 +1,0 @@
-<ion-card>
-  <ion-row align-items-center>
-    <ion-col col-auto>
-      <journal-indicators [showSpecialNeedsIndicator]="isSpecialNeedsSlot()"></journal-indicators>
-    </ion-col>
-    <ion-col col-2>
-      <journal-time [time]="slot.slotDetail.start" [testComplete]="testComplete"></journal-time>
-    </ion-col>
-    <ion-col col-5>
-      <journal-candidate 
-        [name]="slot.booking.candidate.candidateName"
-        [testCategory]="slot.vehicleSlotType"
-        [welshLanguage]="slot.booking.application.welshTest">
-      </journal-candidate>
-    </ion-col>
-    <ion-col col-2>
-       <journal-test-details [testCentreName]="slot.testCentre.centreName"></journal-test-details>
-    </ion-col>
-    <ion-col col-2>
-      <journal-test-outcome [slot]="slot"></journal-test-outcome>
-    </ion-col>
-  </ion-row>
-</ion-card>
