--- src/pages/journal/journal-components/journal-slot/journal-slot.scss
@@ -1,2 +1,0 @@
-journal-slot {
-}
