--- src/pages/journal/journal-components/journal-test-details/journal-test-details.scss
@@ -1,3 +1,0 @@
-journal-test-details {
-
-}
