--- src/pages/journal/journal-components/journal-test-details/journal-test-details.ts
@@ -1,12 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-  selector: 'journal-test-details',
-  templateUrl: 'journal-test-details.html'
-})
-export class JournalTestDetailsComponent {
-  @Input()
-  testCentreName: string;
-
-  constructor() {}
-}
