--- src/pages/journal/journal-components/journal-test-outcome/journal-test-outcome.scss
@@ -1,3 +1,0 @@
-journal-test-outcome {
-
-}
