--- src/pages/journal/journal-components/journal-time/journal-time.html
@@ -1,7 +1,0 @@
-<div>
-    <ion-grid>
-        <ion-row align-items-left>
-            <ion-col><span [ngClass]="{'journal-time-test-complete-text': testComplete}">{{time | date: "HH:mm" }}</span></ion-col>
-        </ion-row>
-    </ion-grid>
-</div>
