--- src/pages/journal/journal-components/journal-time/journal-time.scss
@@ -1,5 +1,0 @@
-journal-time {
-    &.journal-time-test-complete-text {
-        color: map-get($colors-gds, gds-grey-1);
-      }
-}
