--- src/pages/journal/journal-components/journal-time/journal-time.ts
@@ -1,22 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-  selector: 'journal-time',
-  templateUrl: 'journal-time.html'
-})
-export class JournalTimeComponent {
-
-  @Input()
-  time: string;
-
-  @Input()
-  testComplete: boolean;
-
-  ngOnInit() {
-    this.testComplete = true; // TODO pick up testComplete from JSON when available
-  }
-
-  constructor() {
-  }
-
-}
