+++ src/pages/journal/journal-components/test-details/test-details.html
