+++ src/pages/journal/journal-components/test-details/test-details.scss
@@ -1,0 +1,3 @@
+test-details {
+
+}
