+++ src/pages/journal/journal-components/test-details/test-details.ts
@@ -1,0 +1,12 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'test-details',
+  templateUrl: 'test-details.html'
+})
+export class TestDetailsComponent {
+  @Input()
+  testCentreName: string;
+
+  constructor() {}
+}
