+++ src/pages/journal/journal-components/test-outcome/__tests__/test-outcome.spec.ts
@@ -1,0 +1,34 @@
+import { ComponentFixture, async, TestBed } from '@angular/core/testing';
+import { IonicModule } from 'ionic-angular';
+import { TestOutcomeComponent } from '../test-outcome';
+
+
+describe('Test Outcome', () => {
+  let fixture: ComponentFixture<TestOutcomeComponent>;
+  let component: TestOutcomeComponent;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [TestOutcomeComponent],
+      imports: [IonicModule.forRoot(TestOutcomeComponent)],
+      providers: []
+    })
+      .compileComponents()
+      .then(() => {
+        fixture = TestBed.createComponent(TestOutcomeComponent);
+        component = fixture.componentInstance;
+      });
+  }));
+
+  describe('Class', () => {
+    // Unit tests for the components TypeScript class
+    it('should create', () => {
+      expect(component).toBeDefined();
+    });
+  });
+
+  describe('DOM', () => {
+    // Unit tests for the components template
+
+  });
+});
