+++ src/pages/journal/journal-components/test-outcome/test-outcome.scss
@@ -1,0 +1,3 @@
+test-outcome {
+
+}
