+++ src/pages/journal/journal-components/test-outcome/test-outcome.ts
@@ -1,0 +1,18 @@
+import { Component, Input, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'test-outcome',
+  templateUrl: 'test-outcome.html'
+})
+export class TestOutcomeComponent implements OnInit {
+  @Input()
+  slot: any;
+
+  canStartTest: boolean = true;
+  outcome: string = '1'; // todo - get the activity code
+
+  constructor() {}
+
+  ngOnInit() {}
+
+}
