+++ src/pages/journal/journal-components/test-slot/test-slot.scss
@@ -1,0 +1,2 @@
+test-slot {
+}
