+++ src/pages/journal/journal-components/time/time.html
@@ -1,0 +1,7 @@
+<div>
+    <ion-grid>
+        <ion-row align-items-left>
+            <ion-col><span [ngClass]="{'time-test-complete-text': testComplete}">{{time | date: "HH:mm" }}</span></ion-col>
+        </ion-row>
+    </ion-grid>
+</div>
