+++ src/pages/journal/journal-components/time/time.scss
@@ -1,0 +1,5 @@
+time {
+    &.time-test-complete-text {
+        color: map-get($colors-gds, gds-grey-1);
+      }
+}
